+++ src/bpms/bpmn/index.ts
@@ -1,2 +1,3 @@
-export * from "./BpmnProcess";
+export * from "./BpmnProcessInstance";
 export * from "./BpmnEngine";
+export * from "./BpmnProcessPersistency";